2003-12-28  Tim Coleman <tim@timcoleman.com>         * DbParameterBase.cs:                 Correct type of ValueSize method to int
svn path=/trunk/mcs/; revision=21519

+++ mcs/class/System.Data/System.Data.ProviderBase/DbParameterBase.cs
@@ -166,7 +166,7 @@ namespace System.Data.ProviderBase {
 		}
 
 		[MonoTODO]
-		protected virtual byte ValueSize (object value)
+		protected virtual int ValueSize (object value)
 		{
 			throw new NotImplementedException ();
 		}
